# docs/openapi.json
{
  "components": {
    "schemas": {
      "Error": {
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "JSON API behind the Project Divert mobile app and admin tooling.\n\nThis document is generated from the application source by `scripts/generate_openapi.py`: paths and methods come from the Flask URL map, authentication and roles from the `jwt_required` decorator, request fields from the body and query reads in each view, and response codes from the status codes each view returns. `tests/api/test_openapi_contract.py` fails if it falls out of date.\n\nResponse property lists are best-effort: they are read from the serialiser each endpoint uses, so they name the keys you can expect rather than constraining their types.",
    "license": {
      "name": "MIT",
      "url": "https://github.com/dodsy4/projectdivert/blob/main/LICENSE"
    },
    "title": "Project Divert API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/billing/requests": {
      "get": {
        "description": "Requires a bearer token with role: `admin`. Error messages returned by this endpoint: \"Failed to query billing requests\"; \"Invalid billing state filter\".",
        "operationId": "api_admin_list_billing_requests",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "reference",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "request_status",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "filters": {},
                    "items": {},
                    "pagination": {},
                    "summary": {}
                  },
                  "type": "object"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List billing requests (admin)",
        "tags": [
          "Admin / Billing"
        ]
      }
    },
...
    "/api/v1/auth/password-reset/request": {
      "post": {
        "description": "Public endpoint: no bearer token required. Error messages returned by this endpoint: \"A valid email is required\".",
        "operationId": "api_auth_password_reset_request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "email": {}
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Auth password reset request",
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/api/v1/auth/verify/request": {
      "post": {
        "description": "Public endpoint: no bearer token required. Error messages returned by this endpoint: \"A valid email is required\".",
        "operationId": "api_auth_verify_request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "email": {}
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Auth verify request",
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/api/v1/waste-requests": {
      "post": {
        "description": "Requires a bearer token with role: `admin`, `customer`. Error messages returned by this endpoint: \"Failed to create waste request\"; \"Missing required field(s)\"; \"Token missing email claim\"; \"custom_material_type is required when material_type is Other\"; \"match_radius_miles must be a positive number\"; \"scheduled_pickup_at must be in the future\"; \"waste_amount must be a positive number\".",
        "operationId": "api_create_waste_request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "custom_material_type": {},
                  "match_radius_miles": {},
                  "material_type": {},
                  "notes": {},
                  "pickup_address": {},
                  "pickup_city": {},
                  "pickup_county": {},
                  "pickup_postcode": {},
                  "requester_email": {},
                  "requester_name": {},
                  "scheduled_pickup_at": {},
                  "waste_amount": {},
                  "waste_unit": {}
                },
                "required": [
                  "match_radius_miles",
                  "material_type",
                  "pickup_address",
                  "pickup_postcode",
                  "requester_email",
                  "requester_name",
                  "scheduled_pickup_at",
                  "waste_amount",
                  "waste_unit"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "dispatch": {},
                    "drive_time": {},
                    "match": {},
                    "request": {}
                  },
                  "type": "object"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create waste request",
        "tags": [
          "Waste requests"
        ]
      }
    },
    "/api/v1/waste-requests/{request_id}": {
      "get": {
        "description": "Requires a bearer token with role: `admin`, `customer`, `driver`. Error messages returned by this endpoint: \"Forbidden\"; \"Waste request not found\".",
        "operationId": "api_get_waste_request",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "billing": {},
                    "communication_summary": {},
                    "communications": {},
                    "compliance": {},
                    "dispatch": {},
                    "financials": {},
                    "latest_location": {},
                    "match": {},
                    "request": {}
                  },
                  "type": "object"
                }
              }
            },
            "description": "Success"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get waste request",
        "tags": [
          "Waste requests"
        ]
      }
    },
...
    "/api/v1/waste-requests/{request_id}/events": {
      "get": {
        "description": "Requires a bearer token with role: `admin`, `customer`, `driver`. Error messages returned by this endpoint: \"Forbidden\"; \"Waste request not found\".",
        "operationId": "api_stream_waste_request_events",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "last_event_id",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "Success"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Stream waste request events",
        "tags": [
          "Waste requests"
        ]
      }
    },
    "/api/v1/waste-requests/{request_id}/location": {
      "post": {
        "description": "Requires a bearer token with role: `admin`, `driver`. Error messages returned by this endpoint: \"Forbidden\"; \"Waste request not found\"; \"latitude and longitude are required numeric values\".",
        "operationId": "api_create_vehicle_location",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "latitude": {},
                  "longitude": {},
                  "recorded_at": {},
                  "source": {},
                  "vehicle_id": {}
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "location": {}
                  },
                  "type": "object"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create vehicle location",
        "tags": [
          "Waste requests"
        ]
      }
    },
    "/api/v1/waste-requests/{request_id}/location/latest": {
      "get": {
        "description": "Requires a bearer token with role: `admin`, `customer`, `driver`. Error messages returned by this endpoint: \"Forbidden\"; \"No location updates for this request yet\"; \"Waste request not found\".",
        "operationId": "api_get_latest_vehicle_location",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "latest_location": {},
                    "request_id": {},
                    "request_status": {}
                  },
                  "type": "object"
                }
              }
            },
            "description": "Success"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get latest vehicle location",
        "tags": [
          "Waste requests"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "This deployment",
      "url": "/"
    }
  ],
...
}